server/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
google-genai==1.14.0
supabase==2.15.2
python-dotenv==1.1.0
python-multipart
pillow==11.2.1
